examples: created send pattern example that works.
Note that client side a call `yarp::os::Time::delay` is necessary,
without that the publishing thread goes too fast and some data server
side is not received.
Is it a bug? Maybe yes, but it works with a thread
of 1000 hz frequency, in robotics what sensor publish faster than that
frequency?